+++ UVtools.Core/Operations/OperationEditParameters.cs
@@ -6,7 +6,10 @@
  *  of this license document, but changing it is not allowed.
  */
 
+using System.Linq;
+using System.Text;
 using UVtools.Core.FileFormats;
+using UVtools.Core.Objects;
 
 namespace UVtools.Core.Operations
 {
@@ -21,11 +24,37 @@ namespace UVtools.Core.Operations
         public override string Description =>
             "Edits the available print parameters.";
 
-        public override string ConfirmationText => "commit print parameter changes?";
+        public override string ConfirmationText
+        {
+            get
+            {
+                var sb = new StringBuilder();
+                foreach (var modifier in Modifiers)
+                {
+                    if(!modifier.HasChanged) continue;
+                    sb.AppendLine($"{modifier.Name}: {modifier.OldValue} » {modifier.NewValue}");
+                }
+                return $"commit print parameter changes?\n{sb}";
+            }
+        }
 
         public override string ProgressTitle => null;
 
         public override string ProgressAction => null;
+
+        public override StringTag Validate(params object[] parameters)
+        {
+            var sb = new StringBuilder();
+            var changed = Modifiers.Any(modifier => modifier.HasChanged);
+
+            if (!changed)
+            {
+                sb.AppendLine("Nothing changed\nDo some changes or cancel the operation.");
+            }
+
+
+            return new StringTag(sb.ToString());
+        }
 
         public FileFormat.PrintParameterModifier[] Modifiers { get; set; }